+++ src/lib/TradeMemo.ts
@@ -35,6 +35,8 @@ export class TradeMemo extends PricesHolder {
    */
   private y: number;
 
+  private _lock: boolean;
+
   constructor(tradeResult: TradeResult) {
     super();
     this.tradeResult = tradeResult;
@@ -71,6 +73,18 @@ export class TradeMemo extends PricesHolder {
     );
     tradeMemo.prices = tradeMemo.prices || [];
     return tradeMemo;
+  }
+
+  get locked(): boolean {
+    return !!this._lock;
+  }
+
+  lock(): void {
+    this._lock = true;
+  }
+
+  unlock(): void {
+    this._lock = false;
   }
 
   get currentValue(): number {
@@ -148,9 +162,18 @@ export class TradeMemo extends PricesHolder {
   }
 
   profit(): number {
-    return this.tradeResult.soldPrice
-      ? this.tradeResult.gained - this.tradeResult.paid
-      : this.currentPrice * this.tradeResult.quantity - this.tradeResult.paid;
+    if (this.tradeResult.soldPrice) {
+      return this.tradeResult.gained - this.tradeResult.paid;
+    } else {
+      // using lot size quantity to calculate profit,
+      // because if quantity has a fraction that is less than lot size,
+      // that part will not be sold
+      // hence here we're counting only the part that will be sold
+      const qty = this.tradeResult.lotSizeQty || this.tradeResult.quantity;
+      // anticipated sell commission percentage
+      const commission = 1.0001;
+      return this.currentPrice * qty - this.tradeResult.paid * commission;
+    }
   }
 
   /**