+++ src/lib/Types.ts
@@ -136,6 +136,7 @@ export enum Key {
   MAX_PERCENTILE,
   PRICE_MOVE,
   MIN_PERCENTILE,
+  STRENGTH,
 }
 
 export enum Bit {
@@ -163,4 +164,5 @@ export interface PriceChannelData {
   [Key.MAX_PERCENTILE]: number;
   [Key.MIN_PERCENTILE]: number;
   [Key.PRICE_MOVE]: PriceMove;
+  [Key.STRENGTH]: number;
 }