+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# v3.0.3
+
+* Improved current profit percentage accuracy.
+* Improved candidate strength accuracy.
+* Fixed concurrent trades update issues.
+* Fixes UI state not being immediately updated after trade deletion.
+* Added Setting to Hide balances.
+
 # v3.0.2
 
 * Improved stop-limit grid.